requirements.txt
numpy
pandas
torch
torchvision
torchaudio
biopython
